pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Roman_Sku <romanskuratovich08@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
